# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: El Príncipe Mestizo - Citizen journalism platform API
servers:
  - url: /api/v1
    description: Canonical API v1 path
  - url: /api
    description: Legacy (deprecated — será removido en v2)
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /upload/image:
    post:
      operationId: uploadImage
      tags: [upload]
      summary: Upload image to Cloudinary
      responses:
        "200":
          description: Uploaded image URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
      required:
        - id
        - email
        - displayName
        - role
...
    UploadResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url
...